Ispravljene greske u kodu.

+++ Projekat/GlasajBA/GlasajBA/ViewModel/AdministratorViewModel.cs
@@ -202,12 +202,12 @@ namespace GlasajBA.ViewModel
 
         public void obrisiKandidata(Object o)
         {
-            Kandidat x;
+            Kandidat x=null;
             for(int i=0; i<Sistem.KandidatiD.Count; i++)
             {
                 if (Sistem.KandidatiD[i].JMBG1==NoviKandidat.JMBG1)
                 {
-                    x = KandidatiD.Get(i);
+                    x = Sistem.KandidatiD[i];
                     Sistem.KandidatiD.RemoveAt(i);
                     break;
                 }
@@ -217,7 +217,7 @@ namespace GlasajBA.ViewModel
             {
                 if (Sistem.KandidatiE[i].JMBG1 == NoviKandidat.JMBG1)
                 {
-                    x = KandidatiD.Get(i);
+                    x = Sistem.KandidatiD[i];
                     Sistem.KandidatiE.RemoveAt(i);
                     break;
                 }
@@ -227,7 +227,7 @@ namespace GlasajBA.ViewModel
             {
                 if (Sistem.KandidatiK[i].JMBG1 == NoviKandidat.JMBG1)
                 {
-                    x = KandidatiD.Get(i);
+                    x = Sistem.KandidatiD[i];
                     Sistem.KandidatiK.RemoveAt(i);
                     break;
                 }
@@ -237,7 +237,7 @@ namespace GlasajBA.ViewModel
             {
                 if (Sistem.KandidatiO[i].JMBG1 == NoviKandidat.JMBG1)
                 {
-                    x = KandidatiD.Get(i);
+                    x = Sistem.KandidatiD[i];
                     Sistem.KandidatiO.RemoveAt(i);
                     break;
                 }
@@ -247,15 +247,18 @@ namespace GlasajBA.ViewModel
             {
                 if (ListaKandidata[i].JMBG1==NoviKandidat.JMBG1)
                 {
-                    x = KandidatiD.Get(i);
+                    x = Sistem.KandidatiD[i];
                     ListaKandidata.RemoveAt(i);
                     break;
                 }
             }
-            using (var db=new KandidateDBContext())
+            if (x != null)
             {
-                db.Kandidati.Remove(x);
-                db.SaveChanges();
+                using (var db = new KandidateDBContext())
+                {
+                    db.Kandidati.Remove(x);
+                    db.SaveChanges();
+                }
             }
